[NAYB-29] refactor : Item 경로 수정
