+++ src/test/java/com/prgrms/nabmart/domain/review/ReviewTest.java
@@ -6,14 +6,14 @@ import static org.assertj.core.api.Assertions.catchException;
 import com.prgrms.nabmart.domain.category.MainCategory;
 import com.prgrms.nabmart.domain.category.SubCategory;
 import com.prgrms.nabmart.domain.category.fixture.CategoryFixture;
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import com.prgrms.nabmart.domain.item.exception.NotFoundItemException;
 import com.prgrms.nabmart.domain.review.exception.InvalidReviewException;
 import com.prgrms.nabmart.domain.review.support.ReviewFixture;
 import com.prgrms.nabmart.domain.user.User;
 import com.prgrms.nabmart.domain.user.exception.NotFoundUserException;
+import com.prgrms.nabmart.domain.user.support.UserFixture;
 import com.prgrms.nabmart.global.fixture.ItemFixture;
-import com.prgrms.nabmart.global.fixture.UserFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
