+++ src/test/java/com/prgrms/nabmart/domain/review/service/ReviewServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.BDDMockito.then;
 import com.prgrms.nabmart.domain.category.MainCategory;
 import com.prgrms.nabmart.domain.category.SubCategory;
 import com.prgrms.nabmart.domain.category.fixture.CategoryFixture;
-import com.prgrms.nabmart.domain.item.domain.Item;
+import com.prgrms.nabmart.domain.item.Item;
 import com.prgrms.nabmart.domain.item.repository.ItemRepository;
 import com.prgrms.nabmart.domain.review.Review;
 import com.prgrms.nabmart.domain.review.repository.ReviewRepository;
@@ -16,8 +16,8 @@ import com.prgrms.nabmart.domain.review.support.RegisterReviewCommandFixture;
 import com.prgrms.nabmart.domain.review.support.ReviewFixture;
 import com.prgrms.nabmart.domain.user.User;
 import com.prgrms.nabmart.domain.user.repository.UserRepository;
+import com.prgrms.nabmart.domain.user.support.UserFixture;
 import com.prgrms.nabmart.global.fixture.ItemFixture;
-import com.prgrms.nabmart.global.fixture.UserFixture;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
